product uploading

+++ pages/api/products/[id].js
@@ -5,7 +5,9 @@ export default async function handler(req, res) {
   const {
     method,
     query: { id },
+    cookies,
   } = req;
+  const token = cookies.token;
 
   dbConnect();
 
@@ -18,6 +20,9 @@ export default async function handler(req, res) {
     }
   }
   if (method === "PUT") {
+    if (!token || token !== process.env.token) {
+      return res.status(401).json("Not authenticated!");
+    }
     try {
       const product = await Product.findByIdAndUpdate(id);
       res.status(201).json(product);
@@ -27,6 +32,9 @@ export default async function handler(req, res) {
   }
 
   if (method === "DELETE") {
+    if (!token || token !== process.env.token) {
+      return res.status(401).json("Not authenticated!");
+    }
     try {
       await Product.findByIdAndDelete(id);
       res.status(200).json("The product has been deleted");
